backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8
django-filter>=24.0
drf-spectacular>=0.27
reportlab>=4.2
openpyxl>=3.1
numpy>=1.26
scipy>=1.13
pytest>=8.0
pytest-django>=4.8
